LPS-114186 Remove unneeded map
No need to `map()` + `filter()` if you are not going to create new
objects along the way; we can just use a filter instead.

I believe it was written as a `map()` + `filter()` because in a previous
incarnation[0] of the PR, the code used to use `forEach()` + `push()`,
mutating the children along the way, and that was a natural step from
the imperative/mutable style to the functional/immutable style. But once
we removed the mutation, there's no reason not to just use a simple
`filter()`.

[0]: https://github.com/mateomustapic/liferay-portal/pull/9